requirements.txt
# INSTALL REQUIREMENTS FROM VENV
    # To set up venv:
        # MacOS:
            # python3 -m venv <venv-name>
            # source <venv-name>/bin/activate
        # Windows:
            # python -m venv <venv-name>
            # <venv-name>\Scripts\activate
    # To install packages (both systems):
        # pip install -r requirements.txt

charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
Deprecated==1.2.18
et_xmlfile==2.0.0
fonttools==4.60.1
img2pdf==0.6.1
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.7
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pikepdf==9.11.0
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.4
six==1.17.0
tzdata==2025.2
wrapt==1.17.3